+++ src/lib/form/components/button.svelte
@@ -1,7 +1,0 @@
-<script lang="ts">
-    export let type: 'button' | 'submit';
-</script>
-
-<button {type} {...$$restProps} class="app button-component" on:click>
-    <slot>Button</slot>
-</button>