+++ src/lib/form/components/select.svelte
@@ -1,25 +1,32 @@
 <script lang="ts" generics="T">
     import Error from "./error.svelte";
-    export let
-        label: string,
-        value: string | number| null | T,
-        error: string,
-        blank: boolean = false,
-        multiple: boolean = false,
-        disabled: boolean = false;
+    import LabelInfo from "./label-info.svelte";
+    export let label: string;
+    export let value: string | number | null;
+    export let error: string = '';
+    export let blank: boolean = false;
+    export let multiple: boolean = false;
+    export let disabled: boolean = false;
+    const id = `select-${Math.random() * 5}`;
 </script>
 
-<label class="app input-component" class:disabled>
-    <span>{label}</span>
+<div class="app input-component" class:disabled class:info={$$slots.info}>
+    {#if $$slots.info}
+        <LabelInfo {id} {label}>
+            <slot name="info"></slot>
+        </LabelInfo>
+    {:else}
+        <LabelInfo {id} {label} />
+    {/if}
     {#if multiple}
-        <select multiple bind:value {...$$restProps} on:change autocomplete="off" {disabled}>
+        <select {id} multiple bind:value {...$$restProps} on:change autocomplete="off" {disabled}>
             {#if blank}
                 <option value="">Selecione</option>
             {/if}
             <slot />
         </select>
     {:else}
-        <select bind:value {...$$restProps} on:change autocomplete="off" {disabled}>
+        <select {id} bind:value {...$$restProps} on:change autocomplete="off" {disabled}>
             {#if blank}
                 <option value="">Selecione</option>
             {/if}
@@ -28,5 +35,4 @@
     {/if}
     
     <Error name={error} />
-</label>
-    
+</div>