+++ src/lib/form/components/selection-box.svelte
@@ -1,57 +1,64 @@
-<script lang="ts">
+<script lang="ts" generics="T, C extends 'checkbox' | 'radio'">
     import { onMount } from "svelte";
     import Error from "./error.svelte";
     import checkIcon from "./icons/check.svg?raw";
     import circleIcon from "./icons/circle.svg?raw";
-    export let
-        type: 'checkbox' | 'radio',
-        label: string,
-        value: string | number | boolean = '',
-        selected: string | number | boolean | undefined = undefined,
-        checked: boolean = false,
-        group: (string | number | boolean)[] | undefined = undefined,
-        error: string = '',
-        disabled: boolean = false;
+    import LabelInfo from "./label-info.svelte";
+    type GroupType = (C extends 'checkbox' ? T[] : T) | undefined;
+    export let type: C;
+    export let label: string;
+    export let value: T | undefined = undefined;
+    export let checked: boolean = false;
+    export let group: (C extends 'checkbox' ? T[] : T) | undefined = undefined;
+    export let error: string = '';
+    export let disabled: boolean = false;
+    const id = `box-${Math.random() * 5}`;
     function handleGroup(): void
     {
-        if (Array.isArray(group)) {
+        if (Array.isArray(group) && type === 'checkbox') {
             if (checked) {
+                if (!value) throw new TypeError('<SelectionBox>: Ao utilizar bind:group, é obrigatório informar a prop value');
                 group.push(value);
             } else {
-                group = group.filter(item => item !== value);
+                group = group.filter(item => item !== value) as GroupType;
             }
+            group = group;
         }
-        group = group;
     }
     function initGroup(): void
     {
         if (Array.isArray(group)) {
+            if (!value) throw new TypeError('<SelectionBox>: Ao utilizar bind:group, é obrigatório informar a prop value');
             checked = group.includes(value);
         }
     }
     onMount(initGroup);
 </script>
-
-<!-- svelte-ignore a11y-label-has-associated-control -->
-<label class="app input-component selection-box {type}" class:disabled>
-    <span>{label}</span>
+<div class="app input-component selection-box {type}" class:disabled class:info={$$slots.default}>
     {#if type === 'checkbox'}
         {#if Array.isArray(group)}
-            <input type="checkbox" {...$$restProps} bind:checked={checked} {value} autocomplete='off' {disabled} on:change={handleGroup}>
+            <input {id} type="checkbox" {...$$restProps} bind:checked={checked} {value} autocomplete='off' {disabled} on:change={handleGroup}>
         {:else}
-            <input type="checkbox" {...$$restProps} bind:checked={checked} autocomplete='off' {disabled} on:change>
+            <input {id} type="checkbox" {...$$restProps} bind:checked={checked} autocomplete='off' {disabled} on:change>
         {/if}
     {:else if type === 'radio'}
-        <input type="radio" {...$$restProps} bind:group={selected} {value} autocomplete='off' {disabled} on:change>
+        <input {id} type="radio" {...$$restProps} bind:group {value} autocomplete='off' {disabled} on:change>
     {/if}
     {#if type === 'radio'}
-        <span class="box radio" aria-hidden="true">
+        <label for={id} class="box radio" aria-hidden="true">
             {@html circleIcon}
-        </span>
+        </label>
     {:else}
-        <span class="box check" aria-hidden="true">
+        <label for={id} class="box check" aria-hidden="true">
             {@html checkIcon}
-        </span>
+        </label>
     {/if}
-</label>
-<Error name={error} />
+    {#if $$slots.default}
+        <LabelInfo {id} {label}>
+            <slot></slot>
+        </LabelInfo>
+    {:else}
+        <LabelInfo {id} {label} />
+    {/if}
+    <Error name={error} />
+</div>