+++ src/routes/admin/form/+page.svelte
@@ -5,6 +5,9 @@
 
 <main id="app-main">
     <h1>form</h1>
-    <a href="/admin/form/selection-box">-> SelectionBox</a>
+    <a href="/admin/form/selection-box">SelectionBox</a>
+    <br>
+    <br>
+    <a href="/admin/form/info">Teste do botão de informações adicionais</a>
 </main>
 <aside id="app-aside"></aside>