+++ src/routes/redefinir-senha/+page.svelte
@@ -1,0 +1,6 @@
+<script lang="ts">
+    import { title } from "$lib";
+    import { PasswordResetComponent } from "$lib/auth";
+    title('Redefinir senha');
+</script>
+<PasswordResetComponent />